fix(build): harden rustc validation and reorganize tests
