+++ rust/unreal-bindings/Cargo.toml
@@ -16,7 +16,7 @@ majestic-world-core = { path = "../core" }
 
 [dev-dependencies]
 majestic-world-core = { path = "../core", features = ["test-only-unsafe-hooks"] }
+proptest = { version = "1.4", default-features = false, features = ["std"] }
 
 [build-dependencies]
 cbindgen = "^0.27.0"
-proptest = { version = "1.4", default-features = false, features = ["std"] }
